split routes.py
