+++ frontends/onthehouse_flask/onthehouse_flask/endpoints/__init__.py
@@ -7,6 +7,8 @@ from . import common
 from . import image_endpoint
 from . import profile_endpoint
 from . import recipe_endpoint
+from . import register_endpoint
+from . import login_endpoint
 
 site = common.site
 loginmanager = LoginManager(site)
